Improve workflow in R for outputs and plots
Change the workflow in R for observational analysis as follows:
a.	Make a dataframe for all the results
b.	Add one column for the years
c.	Add columns for baseline output and each different intervention scenario (e.g. new_hiv_2025)
d.	Plot every output including the baseline as one line and one intervention as another line in a different colour
e.	Use grid.extra for layout for 2025-2065 (2070 is superfluous)