cumulative outputs r script
Wrote an R script to be able to read in the input parameters and change the general testing to 0 at different years. This time self-testing was reduced to 0. The script then reads in the 16 different output files that were identified as indicators of the epidemic. These are then plotted over time between 2020 and 2070 for each scenario. The cumulative incidence and some other outputs were also plotted for the different scenarios (implementation years) as well as the fold change of these outputs.